+++ include/loot/metadata/message.h
@@ -31,7 +31,6 @@
 #include "loot/enum/message_type.h"
 #include "loot/metadata/conditional_metadata.h"
 #include "loot/metadata/message_content.h"
-#include "loot/struct/simple_message.h"
 
 namespace loot {
 /**
@@ -75,14 +74,6 @@ public:
   LOOT_API explicit Message(const MessageType type,
                             const std::vector<MessageContent>& content,
                             const std::string& condition = "");
-
-  /**
-   * Construct a Message object from a SimpleMessage object.
-   * @param  message
-   *         The SimpleMessage object.
-   * @return A Message object.
-   */
-  LOOT_API explicit Message(const SimpleMessage& message);
 
   /**
    * Get the message type.
@@ -143,34 +134,6 @@ LOOT_API bool operator<=(const Message& lhs, const Message& rhs);
  *          Message object, false otherwise.
  */
 LOOT_API bool operator>=(const Message& lhs, const Message& rhs);
-
-/**
- * Get a given Message as a SimpleMessage given a language.
- * @param  message
- *         The message to convert.
- * @param  language
- *         The preferred language for the message content.
- * @return A SimpleMessage object for the preferred language, or for English
- *         if message text is not available for the given language.
- */
-LOOT_API std::optional<SimpleMessage> ToSimpleMessage(
-    const Message& message,
-    const std::string& language);
-
-/**
- * Get the given messages as simple messages given a language.
- * @param  messages
- *         The messages to convert.
- * @param  language
- *         The preferred language for the message content.
- * @return A vector of SimpleMessage objects for the preferred language, or for
- *         English if message text is not available for the given language. The
- *         order of the input Message objects is preserved, though any messages
- *         without the preferred language or English content will be omitted.
- */
-LOOT_API std::vector<SimpleMessage> ToSimpleMessages(
-    const std::vector<Message>& messages,
-    const std::string& language);
 }
 
 #endif