+++ src/api/metadata/message.cpp
@@ -53,11 +53,6 @@ Message::Message(const MessageType type,
   }
 }
 
-Message::Message(const SimpleMessage& message) :
-    ConditionalMetadata(message.condition),
-    type_(message.type),
-    content_({MessageContent(message.text, message.language)}) {}
-
 MessageType Message::GetType() const { return type_; }
 
 std::vector<MessageContent> Message::GetContent() const { return content_; }
@@ -97,36 +92,4 @@ bool operator>(const Message& lhs, const Message& rhs) { return rhs < lhs; }
 bool operator<=(const Message& lhs, const Message& rhs) { return !(lhs > rhs); }
 
 bool operator>=(const Message& lhs, const Message& rhs) { return !(lhs < rhs); }
-
-std::optional<SimpleMessage> ToSimpleMessage(const Message& message,
-                                             const std::string& language) {
-  auto content = SelectMessageContent(message.GetContent(), language);
-  if (!content.has_value()) {
-    return std::nullopt;
-  }
-
-  SimpleMessage simpleMessage;
-
-  simpleMessage.type = message.GetType();
-  simpleMessage.language = content.value().GetLanguage();
-  simpleMessage.text = content.value().GetText();
-  simpleMessage.condition = message.GetCondition();
-
-  return simpleMessage;
-}
-
-std::vector<SimpleMessage> ToSimpleMessages(
-    const std::vector<Message>& messages,
-    const std::string& language) {
-  std::vector<SimpleMessage> simpleMessages;
-
-  for (const auto& message : messages) {
-    auto simpleMessage = ToSimpleMessage(message, language);
-    if (simpleMessage.has_value()) {
-      simpleMessages.push_back(simpleMessage.value());
-    }
-  }
-
-  return simpleMessages;
-}
 }